third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/blur-while-edit-as-html-expected.txt

Tests that HTML editor hides only when focusing another element


Running: testSetUp

Running: selectNode

Running: testBlurWithoutRelatedTarget
Active element: TEXTAREA
Is editing: true

Running: testBlurWithRelatedTarget
Active element: TEXTAREA
Is editing: false
